% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_to_tdm.R
\name{doc_to_tdm}
\alias{doc_to_tdm}
\title{Converts list of documents to sparse term-document matrix}
\usage{
doc_to_tdm(.out, binary = TRUE)
}
\arguments{
\item{documents}{a list of the output from stm \code{prepDocuments}}
}
\description{
\code{doc_to_tdm} converts list of documents from stm package \code{prepDocuments} function
to a sparse term-document matrix.
}
