% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_to_tdm_fast.R
\name{doc_to_tdm_fast}
\alias{doc_to_tdm_fast}
\title{Converts list of documents to sparse term-document matrix}
\usage{
doc_to_tdm_fast(documents, weights = NULL)
}
\arguments{
\item{documents}{a list of documents from stm \code{prepDocuments}}

\item{weights}{weights to apply to the term-document matrix. must match vocab}
}
\description{
\code{doc_to_tdm_fast} converts list of documents from stm package \code{prepDocuments} function
to a sparse term-document matrix.
}
