% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guest_list.R
\name{guest_list}
\alias{guest_list}
\title{List Available Gifs in Parrot Directory}
\usage{
guest_list()
}
\value{
character vector of available gifs
}
\description{
List Available Gifs in Parrot Directory
}
\examples{
guest_list()
}
