% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{parrot_download}
\alias{parrot_download}
\title{Download Party Parrots for Local Use}
\usage{
parrot_download()
}
\value{
character vector of paths to downloaded files, invisibly
}
\description{
Download Party Parrots for Local Use
}
\examples{
parrot_download()
}
