% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{parrot_download_dir}
\alias{parrot_download_dir}
\title{Parrot Download Directory Path}
\usage{
parrot_download_dir()
}
\value{
path to directory
}
\description{
Parrot Download Directory Path
}
\examples{
parrot_download_dir()
}
