% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{party}
\alias{party}
\title{Print a Party Parrot to the Viewer}
\usage{
party(parrot, ..., height = 30)
}
\arguments{
\item{parrot}{name of parrot gif}

\item{...}{additional arguments to pass to \code{htmltools::div()}}

\item{height}{height of the parrots. Default is 30. Larger may work well for HD parrots.}
}
\value{
HTML version of the parrot
}
\description{
Print a Party Parrot to the Viewer
}
\examples{
party('parrot')
party(c('parrot', 'sadparrot', 'zimbabweparrot'))
}
