% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{party_md}
\alias{party_md}
\title{Return Markdown format of Party Parrots}
\usage{
party_md(parrot, height = 30)
}
\arguments{
\item{parrot}{name of parrot(s) gif}

\item{height}{height of the parrots. Default is 30. Larger may work well for HD parrots.}

\item{...}{additional arguments to pass to \code{htmltools::div()}}
}
\value{
HTML div containing parrots
}
\description{
Return Markdown format of Party Parrots
}
\examples{
party_md('parrot')
party_md(c('parrot', 'sadparrot', 'zimbabweparrot'))
}
