% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_keywords.R
\name{plot_keywords}
\alias{plot_keywords}
\title{Plots keywords}
\usage{
plot_keywords(scores, x_dimension = 1, y_dimension = 2, q_cutoff = 0.9,
  plot_density = FALSE, unstretch = FALSE, print_keywords = T,
  n_words = 12)
}
\arguments{
\item{scores}{list from output of \code{scale_text}}

\item{x_dimension}{integer scalar. dimension for x axis}

\item{y_dimension}{integer scalar. dimension for y axis}

\item{q_cutoff}{numeric (0-1) scalar. proportion of words to plot}

\item{plot_density}{logical scalar. plot both biplot and density}

\item{unstretch}{unstretch space before plotting}

\item{print_keywords}{logical scalar. plot biplot, density, and keywords}

\item{n_words}{number of keywords to print}
}
\description{
\code{plot_keywords} plots keywords for each scaled text dimension
}
