% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_documents.R
\name{score_documents}
\alias{score_documents}
\title{Score documents}
\usage{
score_documents(scores, n_dimensions = 10)
}
\arguments{
\item{scores}{A list from output of \code{scale_text}}

\item{n_dimensions}{An integer scalar, How many dimensions of scaled text to
score.}
}
\description{
\code{score_documents} uses word scores to assign document scores
}
\examples{

\dontrun{
scores <- scale_text(
    meta = out$meta,
    tdm = tdm
    )

document_scores <- score_documents(
    scores = scores, n_dimensions = 10
    )
}

}
\seealso{
\code{\link{scale_text}},
\code{\link{read_word_embeddings}},
\code{\link{get_keywords}}, \code{\link{plot_keywords}},
\code{\link{doc_to_tdm}}
}
