% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_to_tdm.R
\name{doc_to_tdm}
\alias{doc_to_tdm}
\title{Convert list of documents to sparse term-document matrix.}
\usage{
doc_to_tdm(out, binary = TRUE)
}
\arguments{
\item{out}{a list of the output from stm \code{prepDocuments}.}

\item{binary}{if TRUE (default) then only count one occurrence of a word in a document.}
}
\description{
\code{doc_to_tdm} converts list of documents from stm package \code{prepDocuments} function
to a sparse term-document matrix.
}
\examples{

\dontrun{
library(stm)

processed <- textProcessor(
    input_data$text,
    data.frame(input_data),
    removestopwords=T, lowercase=T, stem=F
    )
out <- prepDocuments(
    processed$documents, processed$vocab, processed$meta
    )

tdm <- doc_to_tdm(out)
}

}
