% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{get_keywords}
\alias{get_keywords}
\title{List keywords from \code{scale_text}}
\usage{
get_keywords(scores, n_dimensions, n_words = 15, stretch = 3,
  capture_output = FALSE, pivots_only = TRUE)
}
\arguments{
\item{scores}{List from output of \code{scale_text}.}

\item{n_dimensions}{Integer scalar or vector. How many/which dimensions to
print.}

\item{n_words}{An integer scalar. How many keywords for each dimension.}

\item{stretch}{A scalar. Must be positive, odd integer. Reduce importance of pivot words in
specific (<-- and -->) keywords.}
}
\description{
\code{get_keywords} lists keywords for each scaled text dimension.
}
\examples{
\dontrun{
scores <- scale_text(
    meta=out$meta,
    tdm=tdm
    )

get_keywords(scores, n_dimensions=3, n_words=15)
}

}
\seealso{
\code{\link{read_word_embeddings}},
\code{\link{plot_keywords}},
\code{\link{score_documents}}, \code{\link{doc_to_tdm}}
}
