% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{get_keywords}
\alias{get_keywords}
\title{Lists keywords}
\usage{
get_keywords(scores, n_dimensions, n_words = 15)
}
\arguments{
\item{scores}{list from output of \code{scale_text}}

\item{n_dimensions}{number of dimensions used to produce keywords}

\item{n_words}{how many keywords for each dimension}
}
\description{
\code{get_keywords} lists keywords for each scaled text dimension
}
