% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_keywords.R
\name{plot_keywords}
\alias{plot_keywords}
\title{Plots keywords}
\usage{
plot_keywords(scores, x_dimension = 1, y_dimension = 2,
  q_cutoff = 0.9, plot_density = FALSE, unstretch = FALSE,
  color = FALSE)
}
\arguments{
\item{scores}{List from output of \code{scale_text}.}

\item{x_dimension}{integer scalar. Dimension for x axis.}

\item{y_dimension}{integer scalar. Dimension for y axis.}

\item{q_cutoff}{numeric (0-1) scalar. Proportion of words to plot.}

\item{plot_density}{logical scalar. Plot both biplot and density.}

\item{unstretch}{logical scalar. Move pivot words towrad origin.}

\item{color}{logical scalar. Color words in plot. If vector \code{color} is not provided in \code{scores}, will assign 5 colors using k-means on top 2 through 10 dimensions.}
}
\description{
\code{plot_keywords} plots keywords for each scaled text dimension
}
