% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_word_embeddings.R
\name{read_word_embeddings}
\alias{read_word_embeddings}
\title{Reads vocab from word embedding files quickly and with little memory use}
\usage{
read_word_embeddings(indata, ovefile = NA, ovefile2 = NA,
  wikfile = NA, twifile = NA)
}
\arguments{
\item{indata}{character vector. vocab in data set}

\item{ovefile}{O2M_overlap.txt from meta embeddings}

\item{ovefile2}{O2M_oov.txt from meta embeddings}

\item{wikfile}{glove.6B.300d.txt from Wikipedia embeddings}

\item{twifile}{glove.twitter.27B.200d.txt from Twitter embeddings}
}
\description{
\code{read_word_embeddings} reads vocab from word embedding text files
}
\details{
Wikipedia embeddings: http://nlp.stanford.edu/data/glove.6B.zip
Twitter embeddings: http://nlp.stanford.edu/data/glove.twitter.27B.zip
Meta embeddings: http://www.cis.uni-muenchen.de/~wenpeng/renamed-meta-emb.tar.gz
}
