% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_text.R
\name{scale_text}
\alias{scale_text}
\title{Scale text using canonical pivot analysis}
\usage{
scale_text(tdm, vocab, embedding_values = NULL, embedding_vocab = NULL,
  compress_full = FALSE, n_dimension_compression = 200, n_grams = TRUE,
  meta)
}
\arguments{
\item{tdm}{sparseMatrix. rows are documents, columns are vocab}

\item{vocab}{character vector. must align with columns of tdm}

\item{embedding_values}{matrix of embedding values}

\item{embedding_vocab}{character vector. vocab for rows of embedding values}

\item{compress_full}{logical scalar. TRUE for small data sets}

\item{n_dimension_compression}{integer scalar. how many dimensions of PCA to use}

\item{n_grams}{logical scalar. set to false unless data is extremely clean}

\item{meta}{list of data. output of stm's \code{prepDocuments}. must line up with tdm etc}
}
\description{
\code{scale_text} runs pivoted text scaling
}
