% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_documents.R
\name{score_documents}
\alias{score_documents}
\title{Score documents}
\usage{
score_documents(scores, n_dimensions = 10)
}
\arguments{
\item{scores}{list from output of \code{scale_text}}

\item{n_dimensions}{how many dimensions of scaled text to use}
}
\description{
\code{score_documents} uses word scores to assign document scores
}
